+++ shm/shm.h
@@ -11,8 +11,8 @@
 int sysv_shm_open(size_t size, int flags, int perm);
 void *sysv_shm_attach(int shm_id);
 int sysv_shm_detach(void *addr);
-int sysv_shm_write(int shm_id, void* input, int len, int offset);
-int sysv_shm_read(int shm_id, void* output, int len, int offset);
+int sysv_shm_write(int shm_id, void* input, size_t len, size_t offset);
+int sysv_shm_read(int shm_id, void* output, size_t len, size_t offset);
 size_t sysv_shm_get_size(int shm_id);
 int sysv_shm_lock(int shm_id);
 int sysv_shm_unlock(int shm_id);
